fix(LAP-564): logic show start dialog and tour
